Introduce TimeSeries object.
Abstracts the idea of a sequence of frames into a dictionary.
Automatically caches the frames.
The corresponding functionality has been removed from the interactive
plotting class Iplot.

We may need a way to dump frames in order to avoid exhausting memory
for large simulations.

We may want to implement a plot_spec or plot_item class that
automatically sets up the TimeSeries object.